move some things back to utils, update script imports
